refactor(test): #4361 migrate flake-isolation regressions to child submodule
parent_test_residue (#4267) froze inflight.rs inline test LoC at 5451; the two
fixes plus their deterministic regression tests pushed it to 5626. Per #4269 the
cap must NOT be raised — migrate the tests instead. Moves all four flake-related
tests (the two fixed flaky tests + their two deterministic regression tests)
into a child file src/services/discord/inflight/stall_recovery_tests/
flake_isolation_4361.rs. They reach the shared fixtures via `use super::*`, and
their path still contains `stall_recovery_tests` so the CI `stall_recovery`
filter keeps running them. inflight.rs test LoC drops to 5355 (< 5451, cap
untouched); audit_maintainability --check passes. Regenerates module-inventory
and ARCHITECTURE via generate_inventory_docs.py.
