+++ ARCHITECTURE.md
@@ -471,6 +471,8 @@ src/
 │   │   │   │   └── mod.rs
 │   │   │   ├── save_store/
 │   │   │   │   └── identity_gate.rs
+│   │   │   ├── stall_recovery_tests/
+│   │   │   │   └── flake_isolation_4361.rs
 │   │   │   ├── anchor_repost.rs
 │   │   │   ├── budget.rs
 │   │   │   ├── finalizer_identity.rs
